+++ src/chia_log/handlers/daily_stats/__init__.py
@@ -7,6 +7,7 @@ from ...parsers.harvester_activity_parser import HarvesterActivityMessage
 from ...parsers.wallet_added_coin_parser import WalletAddedCoinMessage
 from ...parsers.partial_parser import PartialMessage
 from ...parsers.block_parser import BlockMessage
+from ...parsers.blockchain_db_parser import BlockchainDbMessage
 
 
 class FinishedSignageConsumer(ABC):
@@ -36,6 +37,12 @@ class BlockConsumer(ABC):
 class WalletAddedCoinConsumer(ABC):
     @abstractmethod
     def consume(self, obj: WalletAddedCoinMessage):
+        pass
+
+
+class BlockchainDbConsumer(ABC):
+    @abstractmethod
+    def consume(self, obj: BlockchainDbMessage):
         pass