+++ src/chia_log/handlers/daily_stats/stats_manager.py
@@ -16,6 +16,7 @@ from . import (
     BlockConsumer,
     WalletAddedCoinConsumer,
     FinishedSignageConsumer,
+    BlockchainDbConsumer,
 )
 from .stat_accumulators.eligible_plots_stats import EligiblePlotsStats
 from .stat_accumulators.wallet_added_coin_stats import WalletAddedCoinStats
@@ -25,11 +26,13 @@ from .stat_accumulators.found_proof_stats import FoundProofStats
 from .stat_accumulators.number_plots_stats import NumberPlotsStats
 from .stat_accumulators.found_partial_stats import FoundPartialStats
 from .stat_accumulators.found_block_stats import FoundBlockStats
+from .stat_accumulators.disk_space_stats import DiskSpaceStats
 from src.chia_log.parsers.wallet_added_coin_parser import WalletAddedCoinMessage
 from src.chia_log.parsers.harvester_activity_parser import HarvesterActivityMessage
 from src.chia_log.parsers.finished_signage_point_parser import FinishedSignagePointMessage
 from src.chia_log.parsers.partial_parser import PartialMessage
 from src.chia_log.parsers.block_parser import BlockMessage
+from src.chia_log.parsers.blockchain_db_parser import BlockchainDbMessage
 from src.notifier.notify_manager import NotifyManager
 from src.notifier import Event, EventType, EventPriority, EventService
 
@@ -59,6 +62,7 @@ class StatsManager:
             NumberPlotsStats(),
             EligiblePlotsStats(),
             SignagePointStats(),
+            DiskSpaceStats(),
         ]
 
         logging.info(
@@ -113,6 +117,14 @@ class StatsManager:
             return
         for stat_acc in self._stat_accumulators:
             if isinstance(stat_acc, FinishedSignageConsumer):
+                for obj in objects:
+                    stat_acc.consume(obj)
+
+    def consume_blockchain_db_messages(self, objects: List[BlockchainDbMessage]):
+        if not self._enable:
+            return
+        for stat_acc in self._stat_accumulators:
+            if isinstance(stat_acc, BlockchainDbConsumer):
                 for obj in objects:
                     stat_acc.consume(obj)